Some more bitstream & dct coef parsing mods.
I'm now at 96.5 fps on a PIII/666, and I have a few more ideas in mind that
should work too, so tomorrow I guess we should break the 100 fps barrier :)

(This is with the -o null output. With x11 display, I dont go above
35 or so fps, but there is not much I can do about this)

walken

+++ trunk/mpeg2dec-livid/libmpeg2/mpeg2_internal.h
@@ -55,68 +55,68 @@
 //state)
 typedef struct picture_s {
     //-- sequence header stuff --
-    uint8_t *intra_quantizer_matrix;
-    uint8_t *non_intra_quantizer_matrix;
+    uint8_t * intra_quantizer_matrix;
+    uint8_t * non_intra_quantizer_matrix;
 
     uint8_t custom_intra_quantization_matrix[64];
     uint8_t custom_non_intra_quantization_matrix[64];
 
     //The width and height of the picture snapped to macroblock units
-    uint32_t coded_picture_width;
-    uint32_t coded_picture_height;
+    int coded_picture_width;
+    int coded_picture_height;
 
     //-- picture header stuff --
 
     //what type of picture this is (I,P,or B) D from MPEG-1 isn't supported
-    uint32_t picture_coding_type;
+    int picture_coding_type;
 	
     //-- picture coding extension stuff --
 	
     //quantization factor for motion vectors
-    uint8_t f_code[2][2];
+    int f_code[2][2];
     //quantization factor for intra dc coefficients
-    uint16_t intra_dc_precision;
+    int intra_dc_precision;
     //bool to indicate all predictions are frame based
-    uint16_t frame_pred_frame_dct;
+    int frame_pred_frame_dct;
     //bool to indicate whether intra blocks have motion vectors 
     // (for concealment)
-    uint16_t concealment_motion_vectors;
+    int concealment_motion_vectors;
     //bit to indicate which quantization table to use
-    uint16_t q_scale_type;
+    int q_scale_type;
     //bool to use different vlc tables
-    uint16_t intra_vlc_format;
+    int intra_vlc_format;
 
     //last macroblock in the picture
-    uint32_t last_mba;
+    int last_mba;
     //width of picture in macroblocks
-    uint32_t mb_width;
+    int mb_width;
 
     //stuff derived from bitstream
 
     //pointer to the zigzag scan we're supposed to be using
-    const uint8_t *scan;
+    const uint8_t * scan;
 
     //Pointer to the current planar frame buffer (Y,Cr,CB)
-    uint8_t *current_frame[3];
+    uint8_t * current_frame[3];
     //storage for reference frames plus a b-frame
-    uint8_t *forward_reference_frame[3];
-    uint8_t *backward_reference_frame[3];
-    uint8_t *throwaway_frame[3];
+    uint8_t * forward_reference_frame[3];
+    uint8_t * backward_reference_frame[3];
+    uint8_t * throwaway_frame[3];
 
     //these things are not needed by the decoder
     //NOTICE : this is a temporary interface, we will build a better one later.
-    uint16_t aspect_ratio_information;
-    uint16_t frame_rate_code;
-    uint16_t progressive_sequence;
-    uint16_t top_field_first;
-    uint16_t repeat_first_field;
-    uint16_t progressive_frame;
+    int aspect_ratio_information;
+    int frame_rate_code;
+    int progressive_sequence;
+    int top_field_first;
+    int repeat_first_field;
+    int progressive_frame;
 } picture_t;
 
 typedef struct motion_s {
     uint8_t * ref_frame[3];
-    int16_t pmv[2][2];
-    uint8_t f_code[2];
+    int pmv[2][2];
+    int f_code[2];
 } motion_t;
 
 // state that is carried from one macroblock to the next inside of a same slice
